pyproject.toml
[project]
name = "shift-bot"
version = "0.1.0"
description = "A simple Discord bot for managing operator shifts"
requires-python = ">=3.8"
dependencies = [
    "discord.py>=2.3.2",
    "python-dotenv>=1.0.0",
    "pytz>=2024.1",
] 